third_party/blink/web_tests/external/wpt/html/canvas/offscreen/layers/2d.layer.malformed-operations.drawImage.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.layer.malformed-operations.drawImage</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.layer.malformed-operations.drawImage</h1>
<p class="desc">Check that exceptions are thrown for operations that are malformed while layers are open.</p>


<script>
test(t => {
  var canvas = new OffscreenCanvas(200, 200);
  var ctx = canvas.getContext('2d');

  const canvas2 = new OffscreenCanvas(200, 200);
  const ctx2 = canvas2.getContext('2d');
  // Shouldn't throw on its own.
  ctx2.drawImage(canvas, 0, 0);
  // Make sure the exception isn't caused by calling the function twice.
  ctx2.drawImage(canvas, 0, 0);
  // Calling again inside a layer should throw.
  ctx.beginLayer();
  assert_throws_dom("InvalidStateError",
                    () => ctx2.drawImage(canvas, 0, 0));

}, "Check that exceptions are thrown for operations that are malformed while layers are open.");
</script>
